Split apart types.
