+++ src/commands/durableTaskScheduler/openTaskHubDashboard.ts
@@ -1,5 +1,5 @@
 import { openUrl, type IActionContext } from "@microsoft/vscode-azext-utils";
-import {type DurableTaskHubResourceModel } from "../../tree/durableTaskScheduler/DurableTaskSchedulerDataBranchProvider";
+import { type DurableTaskHubResourceModel } from "../../tree/durableTaskScheduler/DurableTaskHubResourceModel";
 
 export async function openTaskHubDashboard(_: IActionContext, __: DurableTaskHubResourceModel | undefined): Promise<void> {
     if (!__) {
